fix: [ANDROAPP-6520] refactor uses and location of preference provider to repository to maintain data/ui layer separation
