+++ form/src/main/java/org/dhis2/form/data/FormRepository.kt
@@ -29,5 +29,6 @@ interface FormRepository {
     fun storeFile(id: String, filePath: String?): StoreResult?
     fun areSectionCollapsable(): Boolean
     fun hasLegendSet(dataElementUid: String): Boolean
-    fun getRecordUid(): String
+    fun getListFromPreferences(uid: String): MutableList<String>
+    fun saveListToPreferences(uid: String, list: List<String>)
 }
