+++ form/src/main/java/org/dhis2/form/di/Injector.kt
@@ -63,7 +63,6 @@ object Injector {
             ),
             provideDispatchers(),
             openErrorLocation,
-            providePreferenceProvider(context),
         )
     }
 
@@ -108,6 +107,7 @@ object Injector {
             rulesUtilsProvider = provideRulesUtilsProvider(),
             legendValueProvider = provideLegendValueProvider(context),
             useCompose = useCompose,
+            preferenceProvider = providePreferencesProvider(context),
         )
     }
 
@@ -248,6 +248,10 @@ object Injector {
             defaultWeeklyLabel = context.getString(R.string.week_period_span_default_label),
             defaultBiWeeklyLabel = context.getString(R.string.biweek_period_span_default_label),
         ),
+    )
+
+    private fun providePreferencesProvider(context: Context) = PreferenceProviderImpl(
+        context,
     )
 
     private fun providePreferenceProvider(context: Context) = PreferenceProviderImpl(context)
