+++ form/src/main/java/org/dhis2/form/ui/FormViewModelFactory.kt
@@ -2,7 +2,6 @@ package org.dhis2.form.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import org.dhis2.commons.prefs.PreferenceProvider
 import org.dhis2.commons.viewmodel.DispatcherProvider
 import org.dhis2.form.data.FormRepository
 
@@ -11,14 +10,12 @@ class FormViewModelFactory(
     private val repository: FormRepository,
     private val dispatcher: DispatcherProvider,
     private val openErrorLocation: Boolean,
-    private val preferenceProvider: PreferenceProvider,
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return FormViewModel(
             repository = repository,
             dispatcher = dispatcher,
             openErrorLocation = openErrorLocation,
-            preferenceProvider = preferenceProvider,
         ) as T
     }
 }
